POLIO_1187: move constants out of shared,add shared utils
