+++ plugins/polio/js/src/domains/LQAS-IM/LQAS/LqasAfroOverview/Map/LqasAfroMapPanesContainer.tsx
@@ -10,11 +10,7 @@ import { LoadingSpinner } from 'bluesquare-components';
 import { useDispatch } from 'react-redux';
 import { redirectToReplace } from '../../../../../../../../../hat/assets/js/apps/Iaso/routing/actions';
 import { LQAS_AFRO_MAP_URL } from '../../../../../constants/routes';
-import {
-    COUNTRY,
-    DISTRICT,
-    lqasDistrictColors,
-} from '../../../shared/constants';
+import { COUNTRY, DISTRICT } from '../../../shared/constants';
 import { MapPanes } from '../../../../Campaigns/MapComponent/MapPanes';
 import {
     useAfroMapShapes,
@@ -26,8 +22,8 @@ import { AfroMapParams, Side } from '../types';
 import { LqasAfroTooltip } from './LqasAfroTooltip';
 import { LqasAfroPopup } from './LqasAfroPopUp';
 import { getRound } from '../utils';
-// import { Bounds } from '../../../../../../../../../hat/assets/js/apps/Iaso/utils/map/mapUtils';
 import { LqasAfroOverviewContext } from '../Context/LqasAfroOverviewContext';
+import { lqasDistrictColors } from '../../constants';
 
 const getMainLayerStyle = shape => {
     return lqasDistrictColors[shape.status] ?? defaultShapeStyle;
