+++ plugins/polio/js/src/domains/LQAS-IM/shared/constants.ts
@@ -1,16 +1,13 @@
+import { LqasIMtype } from '../../../constants/types';
 import { OK_COLOR, WARNING_COLOR, FAIL_COLOR } from '../../../styles/constants';
 
-export const LQAS_DATASTORE_URL = '/api/datastore/lqas_';
 export const IM_POC_URL = '/api/polio/imstats/';
 // Including number in the name so it can be used with parseInt for Table sorting.
 // FIXME there should be a cleaner way to do this
 export const IM_PASS = '1imOK';
 export const IM_WARNING = '2imWarning';
 export const IM_FAIL = '3imFail';
-export const LQAS_POC_URL = '/api/polio/lqasstats/';
-export const LQAS_PASS = '1lqasOK';
-export const LQAS_DISQUALIFIED = '2lqasDisqualified';
-export const LQAS_FAIL = '3lqasFail';
+
 export const IN_SCOPE = 'inScope';
 export const paperElevation = 2;
 export const imDistrictColors = {
@@ -47,62 +44,6 @@ export const imDistrictColors = {
     },
 };
 
-export const lqasDistrictColors = {
-    [LQAS_PASS]: {
-        color: '#5e5e5e',
-        weight: '2',
-        opacity: '1',
-        fillColor: OK_COLOR,
-        fillOpacity: 0.8,
-        zIndex: 999,
-    },
-    [LQAS_DISQUALIFIED]: {
-        color: '#5e5e5e',
-        fillColor: WARNING_COLOR,
-        fillOpacity: 0.8,
-        weight: '2',
-        opacity: '1',
-        zIndex: 999,
-    },
-    [LQAS_FAIL]: {
-        color: '#5e5e5e',
-        fillColor: FAIL_COLOR,
-        fillOpacity: 0.8,
-        weight: '2',
-        opacity: '1',
-        zIndex: 999,
-    },
-    [IN_SCOPE]: {
-        color: '#5e5e5e',
-        opacity: '1',
-        fillColor: 'grey',
-        weight: '2',
-        zIndex: 1,
-    },
-};
-
-// keys of reasons for children not fingermarked in LQAS
-export const lqasNfmKeys = [
-    'childabsent',
-    'House_not_visited',
-    'Other',
-    'Vaccinated_but_not_FM',
-    'Non_Compliance',
-    'Child_was_asleep',
-    'Child_is_a_visitor',
-];
-
-// keys for reasons for absence in LQAS
-export const lqasRfaKeys = [
-    'Market',
-    'School',
-    'In_playground',
-    'Farm',
-    'Other',
-    'Travelled',
-    'unknown',
-];
-
 // keys for reasons for absence in IM
 export const imRfaKeys = [
     'Tot_child_Abs_Farm',
@@ -130,12 +71,15 @@ export const imBarColorTresholds = {
     warning: 90,
 };
 
-export const lqasBarColorTresholds = {
-    ok: 80,
-    warning: 50,
-};
-
 export const defaultRounds = [1, 2];
 
 export const DISTRICT = 'district';
 export const COUNTRY = 'country';
+export const LIST = 'list';
+export const MAP = 'map';
+export const LqasIMView = {
+    lqas: 'lqas' as LqasIMtype,
+    imIHH: 'imIHH' as LqasIMtype,
+    imOHH: 'imOHH' as LqasIMtype,
+    imGlobal: 'imGlobal' as LqasIMtype,
+};
